% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{etherscan_balance}
\alias{etherscan_balance}
\title{Get Ether Balance for multiple Accounts}
\usage{
etherscan_balance(accounts, api_key = etherscan_get_api_key())
}
\arguments{
\item{accounts}{a character vector of account addresses}

\item{api_key}{the etherscan api key}
}
\value{
a data frame with two columns, address and balance.
}
\description{
Currently, you can only query up to 20 accounts with one call.
}
