% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{etherscan_ethprice}
\alias{etherscan_ethprice}
\title{Get ETHER LastPrice Price}
\usage{
etherscan_ethprice(api_key = etherscan_get_api_key())
}
\arguments{
\item{api_key}{the etherscan api key}
}
\value{
a list with 4 named entries
}
\description{
Finds the current price in BTC and USD.
}
