% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{etherscan_ethsupply}
\alias{etherscan_ethsupply}
\title{Get Total Supply of Ether}
\usage{
etherscan_ethsupply(api_key = etherscan_get_api_key())
}
\arguments{
\item{api_key}{the etherscan api key}
}
\value{
The total Ether supply in Wei. To get value in Ether divide result by 1000000000000000000
}
\description{
Get Total Supply of Ether
}
\examples{
ether <- etherscan_ethsupply() / Rmpfr::mpfr("1000000000000000000")
}
