% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{etherscan_set_api_key}
\alias{etherscan_set_api_key}
\title{Sets the api key as an environment variable}
\usage{
etherscan_set_api_key(api_key)
}
\arguments{
\item{api_key}{etherscan.io api key}
}
\description{
Sets the api key as an environment variable
}
