% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{etherscan_transactions}
\alias{etherscan_transactions}
\title{Get all 'normal' transactions of an account}
\usage{
etherscan_transactions(account, startblock = 0, endblock = 999999999,
  offset = 1000, page = 1, api_key = etherscan_get_api_key())
}
\arguments{
\item{account}{the account address}

\item{startblock}{the starting block}

\item{endblock}{the endblock}

\item{offset}{max records to return}

\item{page}{page number}

\item{api_key}{the etherscan api key}
}
\description{
Get all 'normal' transactions of an account
}
