% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{to_ether}
\alias{to_ether}
\title{Converts values
Currently only from wei to ether is supported}
\usage{
to_ether(x, from = "wei")
}
\arguments{
\item{x}{a vector of numbers (as string representation is possible)}

\item{from}{from scale}
}
\description{
Converts values
Currently only from wei to ether is supported
}
